CRUD 26-06 noite
Adaptações para melhorar meu código e deixar melhor apresentado.
