+++ css/styles.css
@@ -163,6 +163,41 @@ main .register-link a:hover {
     vertical-align: middle;
 }
 
+/* Main Styles for Portal */
+.portal-main {
+    max-width: 900px;
+    margin: 20px auto;
+    padding: 20px;
+    background-color: #333;
+    border-radius: 5px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+    min-height: 80vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.button-group {
+    display: flex;
+    gap: 20px;
+}
+
+.button-group .button {
+    background-color: #007bff;
+    color: #fff;
+    border: none;
+    padding: 10px 20px;
+    text-decoration: none;
+    border-radius: 5px;
+    cursor: pointer;    
+    text-align: center;
+    height: 80px;
+}
+
+.button-group .button:hover {
+    background-color: #0056b3;
+}
+
 /* Estilo main para o Portal */
 .gerenciar-main {
     max-width: 1000px;
